+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [1.4.15] - Unreleased
 
+### Added
+
+- Functionality for `Boss` rooms
+
 ### Changed
 
 - Bump `s4c` to `0.5.0`
@@ -10,7 +14,7 @@
 - Ensure width of animations is divisible by pixel count
 - Update `configure.ac` to use `-lX11` for `raylib` build on `linux`, `macOS`
 - Avoid drawing control hints over notifications
-- Avoid breaking win_condition on debug floor regen
+- Avoid breaking `win_condition` on debug floor regen
 
 ## [1.4.14] - 2025-12-28